requirements.txt
anthropic
python-dotenv
gspread
google-auth
google-auth-oauthlib
google-api-python-client
supabase
flask
flask-cors